split sharding-orchestration-center to sharding-orchestration-center-api and sharding-orchestration-center-xxx
